Refactor into classes
