+++ packages/sfpowerscripts-cli/src/impl/changelog/CommitUpdater.ts
@@ -1,0 +1,54 @@
+import { Release } from "@dxatscale/sfpowerscripts.core/lib/changelog/interfaces/ReleaseChangelogInterfaces";
+import { Changelog as PackageChangelog } from "@dxatscale/sfpowerscripts.core/lib/changelog/interfaces/GenericChangelogInterfaces";
+import * as fs from "fs-extra";
+
+export default class CommitUpdater {
+
+  constructor(
+    private latestRelease: Release,
+    private artifactsToLatestCommitId: {[p: string]: string},
+    private packagesToChangelogFilePaths: {[p: string]: string}
+  ){}
+
+  /**
+   * Generate commits in latest release, for each artifact
+   * Also sets new latestCommitId for artifacts
+   * @returns
+   */
+  update(): void {
+    for (let artifact of this.latestRelease["artifacts"]) {
+      let packageChangelog: PackageChangelog = JSON.parse(
+        fs.readFileSync(this.packagesToChangelogFilePaths[artifact.name], 'utf8')
+      );
+
+      let indexOfLatestCommitId;
+      if (this.artifactsToLatestCommitId?.[artifact.name]) {
+        indexOfLatestCommitId = packageChangelog["commits"].findIndex( (commit) =>
+          commit["commitId"] === this.artifactsToLatestCommitId[artifact.name]
+        );
+        if (indexOfLatestCommitId === -1) {
+          console.log(`Cannot find commit Id ${this.artifactsToLatestCommitId[artifact.name]} in ${artifact.name} changelog`);
+          console.log("Assuming that there are no changes...");
+          artifact["commits"] = [];
+          continue;
+        }
+      }
+      console.log("indexOfLatestCommitId", indexOfLatestCommitId);
+
+      if (indexOfLatestCommitId > 0) {
+        artifact["commits"] = packageChangelog["commits"].slice(0, indexOfLatestCommitId);
+      } else if (indexOfLatestCommitId === 0) {
+        // Artifact verison has not changed
+        artifact["commits"] = [];
+        // Skip to next artifact
+        continue;
+      } else if (indexOfLatestCommitId === undefined ) {
+        // Artifact was not in previous release
+        artifact["commits"] = packageChangelog["commits"];
+      }
+
+      // Set new latestCommitId
+      artifact["latestCommitId"] = packageChangelog["commits"][0]["commitId"];
+    }
+  }
+}
